exception event refactoring
